Backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
groq
Pillow
pandas
python-dotenv
